email_summarizer/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
joblib
python-dotenv
openai
apache-airflow